+++ lib/iface/include/metashell/iface/engine.hpp
@@ -26,6 +26,7 @@
 #include <metashell/iface/preprocessor_tracer.hpp>
 #include <metashell/iface/type_shell.hpp>
 
+#include <metashell/data/engine_config.hpp>
 #include <metashell/data/engine_name.hpp>
 
 namespace metashell
@@ -40,6 +41,8 @@ namespace metashell
       virtual data::real_engine_name name() const = 0;
 
       virtual data::engine_name display_name() const = 0;
+
+      virtual data::engine_config config() const = 0;
 
       virtual iface::type_shell& type_shell() = 0;
       virtual const iface::type_shell& type_shell() const = 0;